Try generating type definitions.
