+++ build.js
@@ -102,12 +102,11 @@ shell.echo("\n\nCompiling protobuf source");
 shell.exec("yarn protobuild");
 
 shell.echo("\n\nCompiling all source from TypeScript to ES6 JS and removing unit test files");
-shell.exec("./node_modules/typescript/bin/tsc --target ES6 --sourceMap false --module CommonJS --outDir ./dist/src");
+shell.exec("./node_modules/typescript/bin/tsc --declaration --declarationMap --target ES6 --sourceMap false --module CommonJS --outDir ./dist/src");
 shell.exec("find dist -type d -name tests -prune -exec rm -rf {} \\;");
 
 //Copy in various files that we need to deploy as part of our NPM package
 shell.cp("./package.json", "./dist");
-shell.cp("./tenant-security-nodejs.d.ts", "./dist");
 shell.cp("./README.md", "./dist");
 shell.cp("./LICENSE", "./dist");
 shell.cp("-R", "./proto", "./dist");
